crawler/pipeline/requirements.txt
setuptools
wheel
cython
flask
flask-cors
flask-swagger
flask_restful
flask-restplus
requests
elasticsearch
numpy
pika
tqdm
pytz
python-dateutil
urllib3
werkzeug==0.16.1

